refactor ParserDialog to remove duplicated code #2952
