+++ analysis/import/GitLogParser/src/main/kotlin/de/maibornwolff/codecharta/importer/gitlogparser/RepoScanParserDialog.kt
@@ -1,7 +1,6 @@
 package de.maibornwolff.codecharta.importer.gitlogparser
 
 import com.github.kinquirer.KInquirer
-import com.github.kinquirer.components.promptConfirm
 import com.github.kinquirer.components.promptInput
 import de.maibornwolff.codecharta.tools.interactiveparser.ParserDialogInterface
 
@@ -15,28 +14,8 @@ class RepoScanParserDialog {
                 hint = "path/to/repo/root"
             )
 
-            val outputFileName: String = KInquirer.promptInput(
-                message = "What is the name of the output file? If empty, the result will be returned to stdOut",
-                hint = "path/to/output/filename.cc.json"
-            )
-
-            val isCompressed = (outputFileName.isEmpty()) || KInquirer.promptConfirm(
-                message = "Do you want to compress the output file?",
-                default = true
-            )
-
-            val isSilent: Boolean =
-                KInquirer.promptConfirm(message = "Do you want to suppress command line output?", default = false)
-
-            val addAuthor: Boolean =
-                KInquirer.promptConfirm(message = "Do you want to add authors to every file?", default = false)
-
             return listOfNotNull(
                 if (repoPath.isBlank()) null else "--repo-path=$repoPath",
-                "--output-file=$outputFileName",
-                if (isCompressed) null else "--not-compressed",
-                "--silent=$isSilent",
-                "--add-author=$addAuthor"
             )
         }
     }
